Contact dynamic address

+++ database/migrations/2026_05_19_102712_add_image_to_categories_table.php
@@ -11,9 +11,11 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('categories', function (Blueprint $table) {
-            $table->string('image')->nullable();
-        });
+        if (!Schema::hasColumn('categories', 'image')) {
+            Schema::table('categories', function (Blueprint $table) {
+                $table->string('image')->nullable();
+            });
+        }
     }
 
     /**
@@ -21,8 +23,10 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('categories', function (Blueprint $table) {
-            $table->dropColumn('image');
-        });
+        if (Schema::hasColumn('categories', 'image')) {
+            Schema::table('categories', function (Blueprint $table) {
+                $table->dropColumn('image');
+            });
+        }
     }
 };
